# client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CalorieCounter",
    "version": "1.0"
  },
  "paths": {
    "/Ingredients": {
      "get": {
        "tags": [
          "Ingredients"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ingredient"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ingredient"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ingredient"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ingredients"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ingredient"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Ingredient"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Ingredient"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Ingredients/{id}": {
      "get": {
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Ingredients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ingredient"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Ingredient"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Ingredient"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Meal": {
      "get": {
        "tags": [
          "Meal"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Meal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMealDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMealDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMealDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Meal/{id}": {
      "get": {
        "tags": [
          "Meal"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Meal"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meal"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meal"
                }
              }
            }
          }
        }
      }
    },
    "/MealTypes": {
      "get": {
        "tags": [
          "MealTypes"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealType"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealType"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MealTypes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MealType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MealType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/MealTypes/{id}": {
      "get": {
        "tags": [
          "MealTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MealType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealType"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MealTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "newName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "MealTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Recepie": {
      "get": {
        "tags": [
          "Recepie"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recepie"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recepie"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recepie"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recepie"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Recepie"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Recepie"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Recepie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Recepie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Recepie/{id}": {
      "get": {
        "tags": [
          "Recepie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Recepie"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recepie"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recepie"
                }
              }
            }
          }
        }
      }
    },
    "/Recepie/{id}/AddIngredient": {
      "put": {
        "tags": [
          "Recepie"
        ],
        "parameters": [
          {
            "name": "recepieId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ingredientId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Recepie/{id}/RemoveIngredient": {
      "put": {
        "tags": [
          "Recepie"
        ],
        "parameters": [
          {
            "name": "recepieId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ingredientId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateMealDto": {
        "required": [
          "date",
          "mealTypeId",
          "quantity",
          "recepieId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "mealTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "recepieId": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Ingredient": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "calories": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Meal": {
        "required": [
          "date",
          "quantity",
          "recepie",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/MealType"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "recepie": {
            "$ref": "#/components/schemas/Recepie"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MealDto": {
        "required": [
          "recepieName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mealTypeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "recepieName": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MealType": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Recepie": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "ingedients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}